Separate out packet class
