+++ src/main/java/com/dylanpdx/retro64/networking/ClientPayloadHandler.java
@@ -1,9 +1,10 @@
 package com.dylanpdx.retro64.networking;
 
+import com.dylanpdx.retro64.networking.packets.McharPacket;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
 public class ClientPayloadHandler {
-    public static void handleMcharPacket(final Retro64Net.McharPacket data, final IPayloadContext context) {
+    public static void handleMcharPacket(final McharPacket data, final IPayloadContext context) {
 
     }
 }
